CardParts.Cards

Description

Combined (value + suit) card data type and its processing methods

Documentation

data Card

This type represents a card - combination of suit and value.

Constructors

Card

Fields

value :: Value

suit :: Suit

parseCard :: String -> Either String Card

This method gets a string which represents full card combination (value + suit) and returns a Card wrapped with Right. The function uses parseValue and parseSuit methods.

Passed string should have length 2 and have format "{VALUE}{SUIT}".

Examples:

parseCard "2c" = Right Two Clubs
parseCard "As" = Right Ace Spades
parseCard "" = Left "Can't process emtpy string"
parseCard "22c" = Left "Argument length should be 2"
parseCard Zd = Left "There is no broadway card, which could be represented with 'Z'"
parseCard "5f" = Left "There is no card suit marked as 'f'"
parseCard "2x" = Left "There is no card suit marked as 'x'"