CardParts.Values

Description

Card value data type and its processing functions

Documentation

data Value

This type represents card value.

Constructors

Two

Three

Four

Five

Six

Seven

Eight

Nine

Ten

Jack

Queen

King

Ace

parseValue :: Char -> ValueResult

This function gets a char which represents card value and returns a Value wrapped with Maybe.

Char should be a digit between 2 and 9 or an uppercase symbol from [AKQJT] list. Otherwise - Nothing returns.

Examples:

parseValue '2' = Right Two
parseValue 'Q' = Right Queen
parseValue 'z' = Left "There is no broadway card, which could be represented with 'z'"
parseValue '1' = Left "There is no number card with value 1"