Game.Calc

Description

Game calculations and its processing functions

Documentation

data PlayerInput

Constructors

PlayerInput

Fields

username :: String

board :: [Combination]

scoop :: Bool

withFantasy :: Bool

data LineResult

Data container for line result for the user.

Contains data about type of line (position - topmiddlebottom), about points for the combination, total values for combo/bonus points after comparison with others.

data LineType

Constructors

Top

Middle

Bottom

data PlayerCalculations

Container for result of calculations for each player.

Contains data about given player input, description of each line in form of LineResult, some utility bool values, such as isScoop, isNextFantasy. Also has total sum of points for user after comparison. And a debug detailed version of the total (totalDetailed field).

Constructors

PlayerCalculations

Fields

player :: PlayerInput

top :: LineResult

middle :: LineResult

bottom :: LineResult

isScoop :: Bool

isNextFantasy :: Bool

totalDetailed :: (String, [LineCompareResult])

total :: Int

comparePlayers :: PlayerInput -> PlayerInput -> LineCompareResult

Function to compare PlayerInput records and retrieve calculated points. Takes two PlayerInput as input. Returns calculated points in form of LineCompareResult

calcGame :: [PlayerInput] -> [PlayerCalculations]

Main function for game calculation.

Takes a list of 'PlayerInput;, which length should be 3. Return a list of PlayerCalculations.

First of all, function receives all players comparison using collectLinesResults. Then it checks for input conditions: length of input and length of comparison results should be equal and should equal 3.

Then for each of player input function creates PlayerCalculations record.